requirements.txt
Django==2.0.3
django-filter==1.1.0
django-material==1.2.2
django-viewflow==1.2.2
pytz==2018.3
